fix commit service

+++ gateway-service/src/main/java/cs544/model/Comment.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @NotBlank
     private String Name;
 
@@ -38,11 +38,13 @@ public class Comment {
 
 
 
-    public Integer getId() {
+    public Long getId() {
+
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
+
         this.id = id;
     }
